Add GeometryClass to ExportGraphics output
Provide GeometryClass for meshes/lines and meshes/lines inside GeometryParts.
