+++ common/changes/@bentley/bentleyjs-core/add-class-to-export-graphics_2021-09-28-16-47.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@bentley/bentleyjs-core",
+      "comment": "Add GeometryClass to ExportGraphics",
+      "type": "none"
+    }
+  ],
+  "packageName": "@bentley/bentleyjs-core"
+}
