+++ common/changes/@bentley/imodeljs-backend/add-class-to-export-graphics_2021-09-28-16-47.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@bentley/imodeljs-backend",
+      "comment": "Add GeometryClass to ExportGraphics",
+      "type": "none"
+    }
+  ],
+  "packageName": "@bentley/imodeljs-backend"
+}
