+++ core/backend/src/test/standalone/GeometryStream.test.ts
@@ -20,10 +20,7 @@ import {
   MassPropertiesRequestProps, PhysicalElementProps, Placement3d, Placement3dProps, TextString, TextStringProps, ThematicGradientMode,
   ThematicGradientSettings, ViewFlags,
 } from "@bentley/imodeljs-common";
-import {
-  ExportGraphics, ExportGraphicsInfo, ExportGraphicsMeshVisitor, ExportGraphicsOptions, GeometricElement, GeometryPart, LineStyleDefinition,
-  PhysicalObject, Platform, SnapshotDb,
-} from "../../imodeljs-backend";
+import { GeometricElement, GeometryPart, LineStyleDefinition, PhysicalObject, Platform, SnapshotDb } from "../../imodeljs-backend";
 import { IModelTestUtils, Timer } from "../IModelTestUtils";
 
 function assertTrue(expr: boolean): asserts expr {
@@ -2520,132 +2517,6 @@ describe("BRepGeometry", () => {
 
     // Expect exception creating sheet body from invalid loop (imprint error)...
     await expect(imodel.getMassProperties(requestProps)).to.be.rejectedWith(Error, "Solid kernel severe error: 942");
-  });
-});
-
-describe("exportGraphics", () => {
-  let imodel: SnapshotDb;
-
-  before(() => {
-    const seedFileName = IModelTestUtils.resolveAssetFile("CompatibilityTestSeed.bim");
-    const testFileName = IModelTestUtils.prepareOutputFile("GeometryStream", "GeometryStreamTest.bim");
-    imodel = IModelTestUtils.createSnapshotFromSeed(testFileName, seedFileName);
-  });
-
-  after(() => {
-    imodel.close();
-  });
-
-  it("converts to IndexedPolyface", async () => {
-    // Set up element to be placed in iModel
-    const seedElement = imodel.elements.getElement<GeometricElement>("0x1d");
-    assert.exists(seedElement);
-    assert.isTrue(seedElement.federationGuid! === "18eb4650-b074-414f-b961-d9cfaa6c8746");
-
-    const box = Box.createRange(Range3d.create(Point3d.createZero(), Point3d.create(1.0, 1.0, 1.0)), true);
-    assert.isFalse(undefined === box);
-
-    const builder = new GeometryStreamBuilder();
-    builder.appendGeometry(box!);
-
-    const elementProps = createPhysicalElementProps(seedElement, undefined, builder.geometryStream);
-    const testElem = imodel.elements.createElement(elementProps);
-    const newId = imodel.elements.insertElement(testElem);
-    imodel.saveChanges();
-
-    const infos: ExportGraphicsInfo[] = [];
-    const onGraphics = (info: ExportGraphicsInfo) => {
-      infos.push(info);
-    };
-    const exportGraphicsOptions: ExportGraphicsOptions = {
-      elementIdArray: [newId],
-      onGraphics,
-    };
-
-    const exportStatus = imodel.exportGraphics(exportGraphicsOptions);
-    assert.strictEqual(exportStatus, DbResult.BE_SQLITE_OK);
-    assert.strictEqual(infos.length, 1);
-    assert.strictEqual(infos[0].color, ColorDef.white.tbgr);
-    assert.strictEqual(infos[0].mesh.indices.length, 36);
-    assert.strictEqual(infos[0].elementId, newId);
-    const polyface = ExportGraphics.convertToIndexedPolyface(infos[0].mesh);
-    assert.strictEqual(polyface.facetCount, 12);
-    assert.strictEqual(polyface.data.pointCount, 24);
-    assert.strictEqual(polyface.data.normalCount, 24);
-    assert.strictEqual(polyface.data.paramCount, 24);
-  });
-
-  //
-  //
-  //    2---3      6
-  //    | \ |     | \
-  //    0---1    4---5
-  //
-  it("ExportMeshGraphicsVisitor", async () => {
-    const numPoints = 7;
-    const numFacets = 3;
-    const pointData = [0, 0, 0, 1, 0, 0, 0, 2, 0, 1, 2, 0, 2, 0, 0, 4, 0, 0, 3, 2, 0];
-    const paramData = [0, 0, 1, 0, 0, 2, 1, 2, 2, 0, 4, 0, 3, 2];
-    const normalData = new Float32Array(pointData.length);
-    const a0 = 2.0;
-    const a1 = 3.0;
-    const b0 = -2.0;
-    const b1 = 5.0;
-    // make normals functionally related to point coordinates . . . not good normals, but good for tests
-    let paramCursor = 0;
-    for (let i = 0; i < pointData.length; i++) {
-      normalData[i] = a1 * pointData[i] + a0;
-      if ((i + 1) % 3 !== 0)
-        paramData[paramCursor++] = b0 + b1 * pointData[i];
-    }
-    const smallMesh = {
-      points: new Float64Array(pointData),
-      params: new Float32Array(paramData),
-      // normals have one-based index as z ..
-      normals: new Float32Array(normalData),
-      indices: new Int32Array([0, 1, 2, 2, 1, 3, 4, 5, 6]),
-      isTwoSided: true,
-    };
-    const knownArea = 4.0;
-    assert.isTrue(smallMesh.points.length === 3 * numPoints);
-    assert.isTrue(smallMesh.normals.length === 3 * numPoints);
-    assert.isTrue(smallMesh.params.length === 2 * numPoints);
-    assert.isTrue(smallMesh.indices.length === 3 * numFacets);
-    const visitor = ExportGraphicsMeshVisitor.create(smallMesh, 0);
-    assert.isDefined(visitor.paramIndex, "paramIndex defined");
-    assert.isDefined(visitor.paramIndex, "paramIndex defined");
-    let numFacetsA = 0;
-    let indexCursor = 0;
-    let areaSum = 0.0;
-    while (visitor.moveToNextFacet()) {
-      numFacetsA++;
-      assert.isTrue(visitor.point.length === 3);
-      assert.isTrue(smallMesh.indices[indexCursor] === visitor.pointIndex[0]);
-      assert.isTrue(smallMesh.indices[indexCursor + 1] === visitor.pointIndex[1]);
-      assert.isTrue(smallMesh.indices[indexCursor + 2] === visitor.pointIndex[2]);
-      const areaVector = visitor.point.crossProductIndexIndexIndex(0, 1, 2)!;
-      areaSum += areaVector.magnitude() * 0.5;
-      assert.isTrue(smallMesh.indices[indexCursor] === visitor.paramIndex![0]);
-      assert.isTrue(smallMesh.indices[indexCursor + 1] === visitor.paramIndex![1]);
-      assert.isTrue(smallMesh.indices[indexCursor + 2] === visitor.paramIndex![2]);
-      assert.isTrue(smallMesh.indices[indexCursor] === visitor.normalIndex![0]);
-      assert.isTrue(smallMesh.indices[indexCursor + 1] === visitor.normalIndex![1]);
-      assert.isTrue(smallMesh.indices[indexCursor + 2] === visitor.normalIndex![2]);
-      for (let k = 0; k < 3; k++) {
-        const point = visitor.point.getPoint3dAtUncheckedPointIndex(k);
-        const normal = visitor.normal!.getPoint3dAtUncheckedPointIndex(k);
-        const param = visitor.param!.getPoint2dAtUncheckedPointIndex(k);
-        for (let j = 0; j < 3; j++) {
-          assert.isTrue(a0 + a1 * point.at(j) === normal.at(j));
-        }
-        for (let j = 0; j < 2; j++) {
-          assert.isTrue(b0 + b1 * point.at(j) === (j === 0 ? param.x : param.y));
-        }
-      }
-      indexCursor += 3;
-    }
-    assert.isTrue(Math.abs(knownArea - areaSum) < 1.0e-13);
-    assert.isTrue(numFacetsA === numFacets, "facet count");
   });
 });
